13) Smart Content Creation/requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
typing-extensions>=4.5.0